+++ components/Navbar.tsx
@@ -1,7 +1,7 @@
 "use client"
 
 import { useEffect } from 'react'
-import { NavLinks } from '@/constants'
+import { NavLinks } from '@/constants/NavigationLinks'
 import Link from 'next/link';
 import React, { useState } from 'react';
 import { GiHamburgerMenu } from 'react-icons/gi'
