+++ constants/SectionHeading.tsx
@@ -1,0 +1,33 @@
+export const HeroSectionHeading = {
+    header:"Full-Stack Developer",
+    subheader:"Embracing the fusion of artistry and code, I shape digital experiences as a front-end developer. With a passion for design and a drive for innovation, I craft seamless interfaces and striking visuals, elevating the aesthetic essence of the web.",
+}
+
+export const AboutSectionHeading = {
+    overline:"about",
+    header:"me and my services 👌",
+    paragraphs: [
+        { id:1, paragraph:"I'm a natural problem solver, eversince I was a kid I enjoyed building   contraptions with lego, solving puzzles, and tinkering my curiosity away. This not only honed my interest in building things but it also polished my persistence and what led me to pursue this career in the present." 
+        },
+    ]
+}
+
+export const TechnologiesSectionHeading = {
+    overline:"technologies",
+    header:"tools i am most familiar with 💻",
+}
+
+export const WorkExperienceSectionHeading = {
+    overline:"work experience",
+    header:"what have i done?",
+}
+
+export const ContactSectionHeading = {
+    overline:"contact",
+    header:"send me a message :)",
+    subheader:"feel free to contact me anytime! i'll get back to you as soon as I can!",
+}
+
+export const FooterSectionHeading = {
+    copyright:"@ 2023 Andrei Sager. All rights reserved"
+}
